4.5 Recognizing Lexemes

+++ interpreter/src/error.rs
@@ -1,8 +1,8 @@
-pub fn error(line: i32, message: &str) {
+pub fn error(line: usize, message: &str) {
     report(line, "", message);
 }
 
-pub fn report(line: i32, where_: &str, message: &str) {
+pub fn report(line: usize, where_: &str, message: &str) {
     eprintln!("[line {}] Error{}: {}", line, where_, message);
     // had_error = true;
     // TODO: had_error isn't implemented yet. Also, use custom Error type.
